+++ src/Svg.Skia/Svg/SvgExtensions.cs
@@ -5140,75 +5140,6 @@ namespace Svg.Skia
         void Draw(SKCanvas canvas, Attributes ignoreAttributes, DrawableBase? until);
     }
 
-    internal abstract class DrawableContainer : DrawableBase
-    {
-        public readonly List<DrawableBase> ChildrenDrawables;
-
-        protected DrawableContainer()
-            : base()
-        {
-            ChildrenDrawables = new List<DrawableBase>();
-        }
-
-        protected virtual void CreateChildren(SvgElement svgElement, SKRect skOwnerBounds, DrawableBase? parent, Attributes ignoreAttributes)
-        {
-            foreach (var child in svgElement.Children)
-            {
-                var drawable = DrawableFactory.Create(child, skOwnerBounds, parent, ignoreAttributes);
-                if (drawable != null)
-                {
-                    ChildrenDrawables.Add(drawable);
-                    Disposable.Add(drawable);
-                }
-            }
-        }
-
-        protected virtual void CreateTransformedBounds()
-        {
-            foreach (var drawable in ChildrenDrawables)
-            {
-                if (TransformedBounds.IsEmpty)
-                {
-                    TransformedBounds = drawable.TransformedBounds;
-                }
-                else
-                {
-                    if (!drawable.TransformedBounds.IsEmpty)
-                    {
-                        TransformedBounds = SKRect.Union(TransformedBounds, drawable.TransformedBounds);
-                    }
-                }
-            }
-        }
-
-        public override void OnDraw(SKCanvas canvas, Attributes ignoreAttributes, DrawableBase? until)
-        {
-            if (until != null && this == until)
-            {
-                return;
-            }
-
-            foreach (var drawable in ChildrenDrawables)
-            {
-                if (until != null && drawable == until)
-                {
-                    break;
-                }
-                drawable.Draw(canvas, ignoreAttributes, until);
-            }
-        }
-
-        public override void PostProcess()
-        {
-            base.PostProcess();
-
-            foreach (var child in ChildrenDrawables)
-            {
-                child.PostProcess();
-            }
-        }
-    }
-
     internal sealed class MaskDrawable : DrawableContainer
     {
         private MaskDrawable()
